Move SMBus code to new file
