+++ LibreHardwareMonitorLib/Hardware/Motherboard/Lpc/F753XX.cs
@@ -18,21 +18,13 @@ internal class F753XX : ISuperIO
     private readonly byte[] _initialFanPwmControl = new byte[2];
     private byte _initialFanPwmMode;
     private readonly bool[] _restoreDefaultFanPwmControlRequired = new bool[2];
-    private readonly byte _chip_address;
+    SmBusDevice Dev;
 
     public F753XX(ChipSmbus chip, byte addr, ushort smb_addr)
     {
         Chip = (Chip)chip;
-        _chip_address = addr;
-        SMB_ADDRESS = smb_addr;
-        SMBHSTSTS = (ushort)(0 + SMB_ADDRESS);
-        SMBHSTCNT = (ushort)(2 + SMB_ADDRESS);
-        SMBHSTCMD = (ushort)(3 + SMB_ADDRESS);
-        SMBHSTADD = (ushort)(4 + SMB_ADDRESS);
-        SMBHSTDAT0 = (ushort)(5 + SMB_ADDRESS);
-        SMBHSTDAT1 = (ushort)(6 + SMB_ADDRESS);
-        SMBAUXCTL = (ushort)(13 + SMB_ADDRESS);
-
+        Dev = new SmBusDevice(addr, smb_addr);
+        
         Voltages = new float?[4];
         Temperatures = new float?[3];
         Fans = new float?[2];
@@ -110,7 +102,7 @@ internal class F753XX : ISuperIO
         r.AppendLine("SMBus " + GetType().Name);
         r.AppendLine();
         r.Append("Base Address: 0x");
-        r.AppendLine(_chip_address.ToString("X4", CultureInfo.InvariantCulture));
+        r.AppendLine(Dev.ChipAddr.ToString("X4", CultureInfo.InvariantCulture));
         r.AppendLine();
 
         if (!Ring0.WaitSmBusMutex(100))
@@ -204,7 +196,7 @@ internal class F753XX : ISuperIO
     {
         if (_restoreDefaultFanPwmControlRequired[index])
         {
-            _initialFanPwmMode = ReadByte(FAN_MODE_REG);
+            WriteByte(FAN_MODE_REG, _initialFanPwmMode);
 
             if (Chip == (Chip)ChipSmbus.F75387)
                 WriteByte(FAN_PWM_EXP_LSB_REG[index], _initialFanPwmControl[index]);
@@ -215,117 +207,14 @@ internal class F753XX : ISuperIO
         }
     }
 
-    private int CheckPre()
-    {
-        ushort status = Ring0.ReadSmbus(SMBHSTSTS);
-        if ((status & SMBHSTSTS_HOST_BUSY) > 0)
-        {
-            return -1;
-        }
-
-        status &= STATUS_FLAGS;
-        if (status > 0)
-        {
-            Ring0.WriteSmbus(SMBHSTSTS, (byte)status);
-            status = (ushort)(Ring0.ReadSmbus(SMBHSTSTS) & STATUS_FLAGS);
-            if (status > 0)
-            {
-                return -1;
-            }
-        }
-        return 0;
-    }
-
-    private int WaitIntr()
-    {
-        const int maxCount = 1000;
-        int timeout = 0;
-        ushort status;
-        bool val;
-        bool val2;
-
-        do
-        {
-            status = Ring0.ReadSmbus(SMBHSTSTS);
-            val = (status & SMBHSTSTS_HOST_BUSY) > 0;
-            val2 = (status & (STATUS_ERROR_FLAGS | SMBHSTSTS_INTR)) > 0;
-
-        } while ((val || !val2) && timeout++ < maxCount);
-
-        if (timeout > maxCount)
-        {
-            return -1;
-        }
-        return status & (STATUS_ERROR_FLAGS | SMBHSTSTS_INTR);
-    }
-
-    private int CheckPost(int status)
-    {
-        if (status < 0)
-        {
-            Ring0.WriteSmbus(SMBHSTCNT, (byte)(Ring0.ReadSmbus(SMBHSTCNT) | SMBHSTCNT_KILL));
-            Thread.Sleep(1);
-            Ring0.WriteSmbus(SMBHSTCNT, (byte)(Ring0.ReadSmbus(SMBHSTCNT) & (~SMBHSTCNT_KILL)));
-
-            Ring0.WriteSmbus(SMBHSTSTS, (byte)STATUS_FLAGS);
-            return -1;
-        }
-
-        Ring0.WriteSmbus(SMBHSTSTS, (byte)status);
-
-        if ((status & SMBHSTSTS_FAILED) > 0 || (status & SMBHSTSTS_DEV_ERR) > 0 || (status & SMBHSTSTS_BUS_ERR) > 0)
-        {
-            return -1;
-        }
-
-        return 0;
-    }
-
-    private int Transaction(byte xact) // 0x0C WORD_DATA, 0x08 BYTE_DATA
-    {
-        int result = CheckPre();
-        if (result < 0)
-            return result;
-
-        Ring0.WriteSmbus(SMBHSTCNT, (byte)(Ring0.ReadSmbus(SMBHSTCNT) & ~SMBHSTCNT_INTREN));
-        Ring0.WriteSmbus(SMBHSTCNT, (byte)(xact | SMBHSTCNT_START));
-
-        int status = WaitIntr();
-
-        return CheckPost(status);
-    }
-
     private byte ReadByte(byte register)
     {
-        if (SMB_ADDRESS == 0)
-            return 0;
-
-        Ring0.WriteSmbus(SMBHSTADD, (byte)(((_chip_address & 0x7f) << 1) | (SMB_READ & 0x01)));
-        Ring0.WriteSmbus(SMBHSTCMD, register);
-
-        Ring0.WriteSmbus(SMBAUXCTL, (byte)(Ring0.ReadSmbus(SMBAUXCTL) & (~SMBAUXCTL_CRC)));
-
-        Transaction(0x08);
-
-        Ring0.WriteSmbus(SMBAUXCTL, (byte)(Ring0.ReadSmbus(SMBAUXCTL) & ~(SMBAUXCTL_CRC | SMBAUXCTL_E32B)));
-
-        return Ring0.ReadSmbus(SMBHSTDAT0);
+        return Dev.ReadByte(register);
     }
 
     private void WriteByte(byte register, byte value)
     {
-        if (SMB_ADDRESS == 0)
-            return;
-
-        Ring0.WriteSmbus(SMBHSTADD, (byte)(((_chip_address & 0x7f) << 1) | (SMB_WRITE & 0x01)));
-        Ring0.WriteSmbus(SMBHSTCMD, register);
-        Ring0.WriteSmbus(SMBHSTDAT0, value);
-
-        Ring0.WriteSmbus(SMBAUXCTL, (byte)(Ring0.ReadSmbus(SMBAUXCTL) & (~SMBAUXCTL_CRC)));
-
-        Transaction(0x08);
-
-        Ring0.WriteSmbus(SMBAUXCTL, (byte)(Ring0.ReadSmbus(SMBAUXCTL) & ~(SMBAUXCTL_CRC | SMBAUXCTL_E32B)));
+        Dev.WriteByte(register, value);
     }
 
     // ReSharper disable InconsistentNaming
@@ -344,41 +233,6 @@ internal class F753XX : ISuperIO
 
     private readonly byte[] FAN_TACHOMETER_MSB_REG = { 0x16, 0x18 };    // read MSB first!
     private readonly byte[] FAN_TACHOMETER_LSB_REG = { 0x17, 0x19 };
-
-    //SMBus:
-    private readonly ushort SMB_ADDRESS = 0;
-
-    private ushort SMBHSTSTS = 0;
-    private ushort SMBHSTCNT = 0;
-    private ushort SMBHSTCMD = 0;
-    private ushort SMBHSTADD = 0;
-    private ushort SMBHSTDAT0 = 0;
-    private ushort SMBHSTDAT1 = 0;
-    private ushort SMBAUXCTL = 0;
-
-    private const byte SMB_READ = 0x01;
-    private const byte SMB_WRITE = 0x00;
-
-    private static ushort SMBAUXCTL_CRC = (1 << 0);
-    private static ushort SMBAUXCTL_E32B = (1 << 1);
-
-    private static ushort SMBHSTCNT_INTREN = (1 << 0);
-    private static ushort SMBHSTCNT_KILL = (1 << 1);
-    //private static ushort SMBHSTCNT_LAST_BYTE = (1 << 5);
-    private static ushort SMBHSTCNT_START = (1 << 6);
-    //private static ushort SMBHSTCNT_PEC_EN = (1 << 7);
-
-    private static ushort SMBHSTSTS_BYTE_DONE = (1 << 7);
-    //private static ushort SMBHSTSTS_INUSE_STS = (1 << 6);
-    //private static ushort SMBHSTSTS_SMBALERT_STS = (1 << 5);
-    private static ushort SMBHSTSTS_FAILED = (1 << 4);
-    private static ushort SMBHSTSTS_BUS_ERR = (1 << 3);
-    private static ushort SMBHSTSTS_DEV_ERR = (1 << 2);
-    private static ushort SMBHSTSTS_INTR = (1 << 1);
-    private static ushort SMBHSTSTS_HOST_BUSY = (1 << 0);
-
-    private static ushort STATUS_ERROR_FLAGS = (ushort)(SMBHSTSTS_FAILED | SMBHSTSTS_BUS_ERR | SMBHSTSTS_DEV_ERR);
-    private static ushort STATUS_FLAGS = (ushort)(SMBHSTSTS_BYTE_DONE | SMBHSTSTS_INTR | STATUS_ERROR_FLAGS);
 
     // ReSharper restore InconsistentNaming
 #pragma warning restore IDE1006 // Naming Styles
